Resolve fixmes
