+++ src/print/ambiguity.ts
@@ -3,24 +3,9 @@ import type { AstPath, doc, Printer } from 'prettier';
 import { printers as estreePrinters } from 'prettier/plugins/estree.js';
 
 import type { Options } from '../options.js';
+import { flattenDoc } from '../utils/doc.js';
 
 const estreePrinter = estreePrinters['estree'] as Printer<Node | undefined>;
-
-// FIXME: Move
-/**
- *
- */
-export function flattenDoc(doc: doc.builders.Doc): string[] {
-  if (Array.isArray(doc)) {
-    return doc.flatMap(flattenDoc);
-  } else if (typeof doc === 'string') {
-    return [doc];
-  } else if ('contents' in doc) {
-    return flattenDoc(doc.contents);
-  } else {
-    return [];
-  }
-}
 
 /**
  * Search next non EmptyStatement node and set current print, so we can fix it
