+++ src/print/template.ts
@@ -4,7 +4,7 @@ import { doc } from 'prettier';
 import { TEMPLATE_TAG_CLOSE, TEMPLATE_TAG_OPEN } from '../config.js';
 import type { Options } from '../options.js';
 import { getTemplateSingleQuote } from '../options.js';
-import { flattenDoc } from './ambiguity.js';
+import { flattenDoc } from '../utils/doc.js';
 
 const {
   builders: { group, hardline, indent, softline },
